requirements.txt
langchain_community==0.3.24
langchain_core==0.3.64
langchain_ollama==0.3.3
langgraph==0.4.8
python-dotenv==1.1.0
typer==0.16.0
typing_extensions==4.14.0
